Renamed simple to cli.
